move queryMap function to city guide folder
